pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "dwdhandler"
dynamic = ["version"]
requires-python = ">=3.10"

dependencies = [
    "numpy",
    "pandas",
    "seaborn",
    "matplotlib",
    "folium",
    "plotly",
    "scipy",
    "psycopg2",
    "python-dotenv"
]

readme = "README.md"

authors = [
  {name = "Tobias Schad", email = "tobias.schad@googlemail.com"}
]
maintainers = [
    {name = "Tobias Schad", email = "tobias.schad@googlemail.com"}
]

description = "Package to handle data from German Weather Service and some simple plotting routines"

license = { file = "LICENSE" }

keywords = ["dwd", "German Weather Service", "Deutscher Wetterdienst"]

[project.urls]
Repository = "https://github.com/MrTscha/dwdhandler"
Issues = "https://github.com/MrTscha/dwdhandler/issues"

[project.optional-dependencies]
rasterproc = ["pyproj"]

[tool.setuptools.dynamic]
version = {attr = "dwdhandler.__version__"}
